code/balena-api/requirements.txt
balena-sdk
flask
flask_cors
python-dotenv